Get A First Passing Test.
The status route is by far the lightest weight route for the entire
application. It merely formats small JSON body and returns a successful
status code. Since this is a serverless collection of routes, this
status check should be a bit more thorough. It should give developers a
quick window into the health of the system and avoid simple status codes
(coming soon...)