Initial Stub For Warehouse Post.
Spike out a potential implementation around what a Warehouse post would
look like. This doesn't really 'massage' the shipment into a body that
the warehouse could use. Rather it takes the data from the event and
stuffs it straight down to the warehouse. This also is devoid of unit
tests for now... ;(