Utilize Got For HTTP Requests.
GOT is kinda like node-fetch but, with way fewer dependencies and issues.
The GitHub repository is shiney clean and the interface is just the same.
Each individual handler should have its' own package.json if extra dependencies
are required. GOT would be an extra dependency.
This will slow down invocation however, this is an event Lambda only.