Finally Get Tested Queue Function.
The SNS queue is going to send a list of messages through to each lambda.
Due to this, the work that's done in these queue lambdas will need to work
on a list of objects. Even further when using Array functions like .map,
one needs to wrap the result in a Promise.all to resolve correctly.

JS can sometimes be a conveluted mess, like Array functions and async/await.